+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Updated to Holochain `v0.5.2`
+- Updated to new Holochain client version `v0.7.0`
 
 ### Added
 - Exposed `on_signal` from the app websocket in the instrumented websocket.